+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+UNRELEASED
+-------------------------
+* Add "serverIPAddress" field to entries.
+
 version 0.1.0 2017-03-05
 -------------------------
 * Initial release