+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 UNRELEASED
 -------------------------
 * Add "serverIPAddress" field to entries.
+* Set bodySize for requests correctly.
 
 version 0.1.0 2017-03-05
 -------------------------